- Unit-tests adjusted to new layout - setup.py changed to put everything in same package
