+++ tools/mo/automation/package_BOM.txt
@@ -1,6 +1,4 @@
 openvino/tools/__init__.py
-openvino/tools/mo/__init__.py
-openvino/tools/mo/__main__.py
 openvino/tools/mo/analysis/__init__.py
 openvino/tools/mo/analysis/boolean_input.py
 openvino/tools/mo/analysis/inputs.py
@@ -68,8 +66,6 @@ openvino/tools/mo/back/TopKNormalizer.py
 openvino/tools/mo/back/TransposeDFT.py
 openvino/tools/mo/back/TransposeReduceFusing.py
 openvino/tools/mo/back/UselessConcatRemoval.py
-openvino/tools/mo/convert.py
-openvino/tools/mo/convert_impl.py
 openvino/tools/mo/front/__init__.py
 openvino/tools/mo/front/ArgOpsSqueeze.py
 openvino/tools/mo/front/ATenToEmbeddingBag.py
@@ -291,6 +287,7 @@ openvino/tools/mo/front/mxnet/extractors/multibox_prior.py
 openvino/tools/mo/front/mxnet/extractors/relu.py
 openvino/tools/mo/front/mxnet/extractors/scaleshift.py
 openvino/tools/mo/front/mxnet/extractors/slice_axis.py
+openvino/tools/mo/front/mxnet/extractors/str_to.py
 openvino/tools/mo/front/mxnet/extractors/utils.py
 openvino/tools/mo/front/mxnet/eye_ext.py
 openvino/tools/mo/front/mxnet/eye_mx_to_eye.py
@@ -698,13 +695,6 @@ openvino/tools/mo/load/onnx/__init__.py
 openvino/tools/mo/load/onnx/loader.py
 openvino/tools/mo/load/tf/__init__.py
 openvino/tools/mo/load/tf/loader.py
-openvino/tools/mo/main.py
-openvino/tools/mo/main_caffe.py
-openvino/tools/mo/main_kaldi.py
-openvino/tools/mo/main_mxnet.py
-openvino/tools/mo/main_onnx.py
-openvino/tools/mo/main_paddle.py
-openvino/tools/mo/main_tf.py
 openvino/tools/mo/middle/__init__.py
 openvino/tools/mo/middle/AddFakeQuantizeFuse.py
 openvino/tools/mo/middle/AddIsCyclicAttribute.py
@@ -824,19 +814,6 @@ openvino/tools/mo/middle/UnsqueezeTileReshapeBlockToInterpolate.py
 openvino/tools/mo/middle/UpsampleToResample.py
 openvino/tools/mo/middle/UselessMerge.py
 openvino/tools/mo/middle/UselessSplitEraser.py
-openvino/tools/mo/mo.py
-openvino/tools/mo/mo_caffe.py
-openvino/tools/mo/mo_kaldi.py
-openvino/tools/mo/mo_mxnet.py
-openvino/tools/mo/mo_onnx.py
-openvino/tools/mo/mo_paddle.py
-openvino/tools/mo/mo_tf.py
-openvino/tools/mo/moc_frontend/__init__.py
-openvino/tools/mo/moc_frontend/analysis.py
-openvino/tools/mo/moc_frontend/check_config.py
-openvino/tools/mo/moc_frontend/extractor.py
-openvino/tools/mo/moc_frontend/pipeline.py
-openvino/tools/mo/moc_frontend/serialize.py
 openvino/tools/mo/ops/__init__.py
 openvino/tools/mo/ops/activation.py
 openvino/tools/mo/ops/activation_ops.py
@@ -1019,22 +996,15 @@ openvino/tools/mo/ops/upsample.py
 openvino/tools/mo/pipeline/__init__.py
 openvino/tools/mo/pipeline/common.py
 openvino/tools/mo/pipeline/unified.py
-openvino/tools/mo/subprocess_main.py
 openvino/tools/mo/utils/__init__.py
 openvino/tools/mo/utils/broadcasting.py
-openvino/tools/mo/utils/check_ie_bindings.py
-openvino/tools/mo/utils/check_mo_import.py
 openvino/tools/mo/utils/class_registration.py
 openvino/tools/mo/utils/cli_parser.py
 openvino/tools/mo/utils/custom_replacement_config.py
 openvino/tools/mo/utils/dsu.py
 openvino/tools/mo/utils/error.py
-openvino/tools/mo/utils/find_ie_version.py
 openvino/tools/mo/utils/find_inputs.py
-openvino/tools/mo/utils/get_ov_update_message.py
 openvino/tools/mo/utils/graph.py
-openvino/tools/mo/utils/guess_framework.py
-openvino/tools/mo/utils/ie_version.py
 openvino/tools/mo/utils/import_extensions.py
 openvino/tools/mo/utils/ir_engine/__init__.py
 openvino/tools/mo/utils/ir_engine/compare_graphs.py
@@ -1088,10 +1058,7 @@ openvino/tools/mo/utils/replacement_pattern.py
 openvino/tools/mo/utils/runtime_info.py
 openvino/tools/mo/utils/shape.py
 openvino/tools/mo/utils/simple_proto_parser.py
-openvino/tools/mo/utils/str_to.py
 openvino/tools/mo/utils/summarize_graph.py
-openvino/tools/mo/utils/telemetry_params.py
-openvino/tools/mo/utils/telemetry_stub.py
 openvino/tools/mo/utils/telemetry_utils.py
 openvino/tools/mo/utils/tensorboard_util.py
 openvino/tools/mo/utils/type_utils.py
@@ -1099,3 +1066,49 @@ openvino/tools/mo/utils/unsupported_ops.py
 openvino/tools/mo/utils/utils.py
 openvino/tools/mo/utils/version.py
 openvino/tools/mo/utils/versions_checker.py
+openvino/tools/mo_lite/__init__.py
+openvino/tools/mo_lite/__main__.py
+openvino/tools/mo_lite/convert.py
+openvino/tools/mo_lite/convert_impl.py
+openvino/tools/mo_lite/main.py
+openvino/tools/mo_lite/main_caffe.py
+openvino/tools/mo_lite/main_kaldi.py
+openvino/tools/mo_lite/main_mxnet.py
+openvino/tools/mo_lite/main_onnx.py
+openvino/tools/mo_lite/main_paddle.py
+openvino/tools/mo_lite/main_tf.py
+openvino/tools/mo_lite/mo.py
+openvino/tools/mo_lite/mo_caffe.py
+openvino/tools/mo_lite/mo_kaldi.py
+openvino/tools/mo_lite/mo_mxnet.py
+openvino/tools/mo_lite/mo_onnx.py
+openvino/tools/mo_lite/mo_paddle.py
+openvino/tools/mo_lite/mo_tf.py
+openvino/tools/mo_lite/moc_frontend/__init__.py
+openvino/tools/mo_lite/moc_frontend/analysis.py
+openvino/tools/mo_lite/moc_frontend/check_config.py
+openvino/tools/mo_lite/moc_frontend/extractor.py
+openvino/tools/mo_lite/moc_frontend/pipeline.py
+openvino/tools/mo_lite/moc_frontend/serialize.py
+openvino/tools/mo_lite/subprocess_main.py
+openvino/tools/mo_lite/utils/__init__.py
+openvino/tools/mo_lite/utils/check_ie_bindings.py
+openvino/tools/mo_lite/utils/check_mo_dev_import.py
+openvino/tools/mo_lite/utils/check_mo_import.py
+openvino/tools/mo_lite/utils/clean_utils.py
+openvino/tools/mo_lite/utils/cli_parser.py
+openvino/tools/mo_lite/utils/error.py
+openvino/tools/mo_lite/utils/find_ie_version.py
+openvino/tools/mo_lite/utils/get_ov_update_message.py
+openvino/tools/mo_lite/utils/guess_framework.py
+openvino/tools/mo_lite/utils/ie_version.py
+openvino/tools/mo_lite/utils/logger.py
+openvino/tools/mo_lite/utils/offline_transformations.py
+openvino/tools/mo_lite/utils/preprocessing.py
+openvino/tools/mo_lite/utils/telemetry_params.py
+openvino/tools/mo_lite/utils/telemetry_stub.py
+openvino/tools/mo_lite/utils/telemetry_utils.py
+openvino/tools/mo_lite/utils/type_utils.py
+openvino/tools/mo_lite/utils/utils.py
+openvino/tools/mo_lite/utils/version.py
+openvino/tools/mo_lite/utils/versions_checker.py
