+++ tools/mo/openvino/tools/mo/utils/cli_parser.py
@@ -1,0 +1,4 @@
+# Copyright (C) 2018-2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino.tools.mo_lite.utils.cli_parser import *
