+++ tools/mo/openvino/tools/mo/utils/versions_checker.py
@@ -1,0 +1,4 @@
+# Copyright (C) 2018-2023 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
+from openvino.tools.mo_lite.utils.versions_checker import *
