+++ tools/mo/openvino/tools/mo_lite/__main__.py
@@ -1,7 +1,7 @@
 # Copyright (C) 2018-2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from openvino.tools.mo.subprocess_main import subprocess_main
-from openvino.tools.mo.utils.telemetry_utils import init_mo_telemetry
+from openvino.tools.mo_lite.subprocess_main import subprocess_main
+from openvino.tools.mo_lite.utils.telemetry_utils import init_mo_telemetry
 init_mo_telemetry()
 subprocess_main(framework=None)
