+++ tools/mo/setup.py
@@ -19,7 +19,7 @@ from setuptools import setup, find_namespace_packages
 from setuptools.command.build_py import build_py
 from setuptools.command.install import install
 
-prefix = 'openvino/tools/mo/'
+prefix = 'openvino/tools/mo_lite/'
 SETUP_DIR = Path(__file__).resolve().parent / Path(prefix)
 
 
@@ -62,9 +62,13 @@ class BuildCmd(build_py):
             for (pkg, module, filename) in modules
         ]
 
+lite_packages = find_namespace_packages(prefix[:-1])
+lite_packages = [prefix.replace('/', '.') + p for p in lite_packages]
 
-packages = find_namespace_packages(prefix[:-1])
-packages = [prefix.replace('/', '.') + p for p in packages]
+dev_packages = find_namespace_packages(prefix[:-6])
+dev_packages = [prefix[:-6].replace('/', '.') + '.' + p for p in dev_packages]
+
+packages = lite_packages + dev_packages
 
 setup(
     name='openvino-mo',
@@ -80,7 +84,7 @@ setup(
     },
     entry_points={
         'console_scripts': [
-            'mo = openvino.tools.mo.__main__:main',
+            'mo = openvino.tools.mo_lite.__main__:main',
         ],
     },
     package_data={
