+++ tools/mo/unit_tests/mo/bom_test.py
@@ -39,7 +39,7 @@ class TestBOMFile(unittest.TestCase):
                 cls.existing_files = [name.rstrip() for name in bom_file.readlines()]
 
         # dirs_to_search is the root directory where MO is located, 'openvino_project_root/tools/mo/openvino/tools'
-        cls.dirs_to_search = os.path.normpath(get_mo_root_dir() + '/mo/')
+        cls.dirs_to_search = [os.path.normpath(get_mo_root_dir() + '/mo/'), os.path.normpath(get_mo_root_dir() + '/mo_lite/')]
         cls.prefix = os.path.normpath(get_mo_root_dir() + '../../../')  # prefix which is used in BOM file
         cls.expected_header = [re.compile(pattern) for pattern in [
             r'^# Copyright \([cC]\) [0-9\-]+ Intel Corporation$',
@@ -48,7 +48,7 @@ class TestBOMFile(unittest.TestCase):
 
     def test_bom_file(self):
         missing_files = list()
-        for src_dir in [self.dirs_to_search]:
+        for src_dir in self.dirs_to_search:
             if not os.path.isdir(src_dir):
                 continue
             for root, dirs, files in os.walk(src_dir):
@@ -66,7 +66,7 @@ class TestBOMFile(unittest.TestCase):
 
         if len(missing_files) != 0:
             print("Missing files:")
-            for f in missing_files:
+            for f in sorted(missing_files):
                 print(f.replace('\\', '/'))
         self.assertTrue(not len(missing_files), '{} files missed in BOM'.format(len(missing_files)))
 
@@ -94,7 +94,7 @@ class TestBOMFile(unittest.TestCase):
 
     def test_missed_intel_header(self):
         missing_files = list()
-        for src_dir in [self.dirs_to_search]:
+        for src_dir in self.dirs_to_search:
             if not os.path.isdir(src_dir):
                 continue
             for root, dirs, files in os.walk(src_dir):
