+++ tools/mo/unit_tests/mo/convert/meta_data_test.py
@@ -9,9 +9,8 @@ from generator import generator
 from openvino.runtime import get_version as get_rt_version
 from openvino.runtime import serialize
 
-from openvino.tools.mo import convert_model
-from openvino.tools.mo.utils import import_extensions
-from openvino.tools.mo.utils.version import get_version
+from openvino.tools.mo_lite import convert_model
+from openvino.tools.mo_lite.utils.version import get_version
 from unit_tests.mo.unit_test_with_mocked_telemetry import UnitTestWithMockedTelemetry
 from utils import save_to_onnx
