+++ tools/mo/unit_tests/mo/convert/meta_data_test_actual.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from openvino.runtime import get_version as get_rt_version
 from openvino.runtime import serialize
-from openvino.tools.mo import convert_model
+from openvino.tools.mo_lite import convert_model
 from openvino.tools.mo.utils.ir_reader.restore_graph import restore_graph_from_ir, save_restored_graph
 from openvino.tools.mo.utils.version import get_version
