+++ tools/mo/unit_tests/mo/frontend_ngraph_test.py
@@ -8,13 +8,13 @@ import sys
 import unittest
 from unittest.mock import patch
 
-from openvino.tools.mo.subprocess_main import setup_env, subprocess_main
+from openvino.tools.mo_lite.subprocess_main import setup_env, subprocess_main
 
 import pytest
 
 
 class TestNoInferenceEngine(unittest.TestCase):
-    @patch('openvino.tools.mo.utils.find_ie_version.find_ie_version')
+    @patch('openvino.tools.mo_lite.utils.find_ie_version.find_ie_version')
     def test_no_ie_ngraph(self, mock_find):
         mock_find.return_value = False
         with pytest.raises(SystemExit) as e, self.assertLogs(log.getLogger(), level="ERROR") as cm:
