+++ tools/mo/unit_tests/mo/frontend_ngraph_test_actual.py
@@ -16,7 +16,7 @@ import pytest
 mock_available = True
 try:
     # pylint: disable=no-name-in-module,import-error
-    from openvino.tools.mo.main import main
+    from openvino.tools.mo_lite.main import main
 
     # pylint: disable=no-name-in-module,import-error
     from mock_mo_python_api import get_frontend_statistic, get_model_statistic, \
