+++ tools/mo/unit_tests/mo/utils/args_to_string_test.py
@@ -4,8 +4,8 @@
 import numpy as np
 from openvino.runtime import Layout, PartialShape, Dimension, Shape, Type
 
-from openvino.tools.mo import InputCutInfo, LayoutMap
-from openvino.tools.mo.utils.cli_parser import input_to_str, mean_scale_value_to_str, \
+from openvino.tools.mo_lite import InputCutInfo, LayoutMap
+from openvino.tools.mo_lite.utils.cli_parser import input_to_str, mean_scale_value_to_str, \
     transform_param_to_str, input_shape_to_str, str_list_to_str, source_target_layout_to_str, layout_param_to_str
 from unit_tests.mo.unit_test_with_mocked_telemetry import UnitTestWithMockedTelemetry
