+++ tools/mo/unit_tests/mo/utils/convert_impl_tmp_irs_cleanup_test_actual.py
@@ -5,8 +5,8 @@ import os
 import unittest
 from unittest.mock import patch
 
-from openvino.tools.mo.convert import convert_model
-from openvino.tools.mo.utils.error import Error
+from openvino.tools.mo_lite.convert import convert_model
+from openvino.tools.mo_lite.utils.error import Error
 
 
 class TestConvertImplTmpIrsCleanup(unittest.TestCase):
@@ -22,7 +22,7 @@ class TestConvertImplTmpIrsCleanup(unittest.TestCase):
         return False
 
     def test_tmp_irs_cleanup_convert_impl_1(self):
-        with patch("openvino.tools.mo.back.offline_transformations.apply_offline_transformations") as emit_ir_func:
+        with patch("openvino.tools.mo_lite.utils.offline_transformations.apply_offline_transformations") as emit_ir_func:
             emit_ir_func.side_effect = Error('offline transformations step has failed')
 
             params = {'input_model': self.test_model_file, 'input_model_is_text': True, 'input': 'x[3],y[1 3]',
@@ -40,7 +40,7 @@ class TestConvertImplTmpIrsCleanup(unittest.TestCase):
             self.assertFalse(self.are_tmp_files_left(self.test_model_file))
 
     def test_tmp_irs_cleanup_convert_impl_3(self):
-        with patch("openvino.tools.mo.convert_impl.read_model") as emit_ir_func:
+        with patch("openvino.tools.mo_lite.convert_impl.read_model") as emit_ir_func:
             emit_ir_func.side_effect = Exception('FEM read_model has failed')
 
             params = {'input_model': self.test_model_file, 'input_model_is_text': True, 'input': 'x[3],y[1 3]',
