+++ tools/mo/unit_tests/mo/utils/freeze_placeholder_test.py
@@ -16,7 +16,7 @@ from openvino.frontend import (
     FrontEnd,
 )  # pylint: disable=no-name-in-module,import-error
 from openvino.runtime import Core
-from openvino.tools.mo.convert_impl import prepare_ir
+from openvino.tools.mo_lite.convert_impl import prepare_ir
 
 
 def base_args_config(use_legacy_fe: bool = None, use_new_fe: bool = None):
@@ -52,7 +52,7 @@ def base_args_config(use_legacy_fe: bool = None, use_new_fe: bool = None):
 try:
     import openvino_telemetry as tm
 except ImportError:
-    import openvino.tools.mo.utils.telemetry_stub as tm
+    import openvino.tools.mo_lite.utils.telemetry_stub as tm
 
 
 def get_test_default_frontends():
@@ -156,7 +156,7 @@ class TestMoFreezePlaceholder(unittest.TestCase):
     )
     def test_freeze_placeholder_with_value_onnx_fe(self, input_freezing_value, use_new_fe, inputs, expected,
                                                    dtype=None):
-        with patch("openvino.tools.mo.convert_impl.get_default_frontends") as default_fe:
+        with patch("openvino.tools.mo_lite.convert_impl.get_default_frontends") as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_new_fe=use_new_fe)
             args.input_model = "test_model.onnx"
@@ -223,7 +223,7 @@ class TestMoFreezePlaceholder(unittest.TestCase):
         ],
     )
     def test_freeze_placeholder_with_value_mul(self, input_freezing_value, use_new_fe, inputs, expected, dtype=None):
-        with patch("openvino.tools.mo.convert_impl.get_default_frontends") as default_fe:
+        with patch("openvino.tools.mo_lite.convert_impl.get_default_frontends") as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_new_fe=use_new_fe)
             args.input_model = "test_model_2.onnx"
@@ -253,7 +253,7 @@ class TestMoFreezePlaceholder(unittest.TestCase):
     )
     def test_value_without_type(self, input_freezing_value, use_new_fe, inputs, expected,
                                 dtype=None):
-        with patch("openvino.tools.mo.convert_impl.get_default_frontends") as default_fe:
+        with patch("openvino.tools.mo_lite.convert_impl.get_default_frontends") as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_new_fe=use_new_fe)
             args.input_model = "test_model_2.onnx"
@@ -283,7 +283,7 @@ class TestMoFreezePlaceholder(unittest.TestCase):
     )
     def test_value_without_type_int32(self, input_freezing_value, use_new_fe, inputs, expected,
                                       dtype=None):
-        with patch("openvino.tools.mo.convert_impl.get_default_frontends") as default_fe:
+        with patch("openvino.tools.mo_lite.convert_impl.get_default_frontends") as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_new_fe=use_new_fe)
             args.input_model = "test_model_int.onnx"
