+++ tools/mo/unit_tests/mo/utils/mo_fallback_test_actual.py
@@ -5,7 +5,7 @@ import unittest
 from unittest.mock import patch, Mock
 
 import openvino
-from openvino.tools.mo.convert_impl import prepare_ir
+from openvino.tools.mo_lite.convert_impl import prepare_ir
 from openvino.tools.mo.utils.error import Error
 from openvino.frontend import FrontEndManager, FrontEnd # pylint: disable=no-name-in-module,import-error
 from onnx.helper import make_graph, make_model, make_tensor_value_info
@@ -21,7 +21,7 @@ from generator import generator, generate
 try:
     import openvino_telemetry as tm
 except ImportError:
-    import openvino.tools.mo.utils.telemetry_stub as tm
+    import openvino.tools.mo_lite.utils.telemetry_stub as tm
 
 def base_args_config(use_legacy_fe:bool=None, use_new_fe:bool=None):
     args = argparse.Namespace()
@@ -150,7 +150,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, None, None, 'onnx_frontend', None),
     ])
     def test_fallback_if_extension_specified(self, extension, use_legacy, use_new_fe, conversion_method, fallback_reason):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_legacy, use_new_fe)
             args.extensions = extension
@@ -174,7 +174,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, True, 'onnx_frontend'),
     ])
     def test_fallback_if_new_extension_specified(self, use_legacy, use_new_fe, conversion_method):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_legacy, use_new_fe)
             args.extensions = ['onnx_fe_ext.so']
@@ -193,7 +193,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, True, 'onnx_frontend'),
     ])
     def test_fallback_if_two_new_extension_specified(self, use_legacy, use_new_fe, conversion_method):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_legacy, use_new_fe)
             args.extensions = ['onnx_fe_ext.so', 'onnx_fe_ext_2.so']
@@ -214,7 +214,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, None, None, 'onnx_frontend', None),
     ])
     def test_fallback_if_tranformations_config_specified(self, trans_config, use_legacy, use_new_fe, expected_path, fallback_reason):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_legacy, use_new_fe)
             args.input_model = "test_model.onnx"
@@ -237,7 +237,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, None, True, 'onnx_frontend', None),
     ])
     def test_fallback_if_both_extension_and_trans_config_specified(self, extension, trans_config, use_new_fe, expected_path, fallback_reason):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = get_test_default_frontends()
             args = base_args_config(use_new_fe=use_new_fe)
             args.extensions = extension
@@ -259,7 +259,7 @@ class TestMoFallback(unittest.TestCase):
                 (None, None, True, 'onnx_frontend'),
     ])
     def test_fallback_if_legacy_set_as_default(self, trans_config, use_legacy, use_new_fe, expected_path):
-        with patch('openvino.tools.mo.convert_impl.get_default_frontends') as default_fe:
+        with patch('openvino.tools.mo_lite.convert_impl.get_default_frontends') as default_fe:
             default_fe.return_value = {'onnx': 'legacy', 'tf': 'legacy'}
             args = base_args_config(use_legacy, use_new_fe)
             args.input_model = "test_model.onnx"
